% Generated by roxygen2 (4.0.1): do not edit by hand
\name{AddAPIKey}
\alias{AddAPIKey}
\title{Function is used to add the API key to the params that will be sent on
every call to OpenSecrets.org; an exception will be thrown if this value is
null.}
\usage{
AddAPIKey(params)
}
\arguments{
\item{params}{Any parameter pertaining to a specific web method invocation.}
}
\description{
Function is used to add the API key to the params that will be sent on
every call to OpenSecrets.org; an exception will be thrown if this value is
null.
}

