% Generated by roxygen2 (4.0.1): do not edit by hand
\name{AddKeyValuePair}
\alias{AddKeyValuePair}
\title{Function is used to add the key and value to the params that will be sent on
every call to OpenSecrets.org; an exception will be thrown if either of
these parameters is null.}
\usage{
AddKeyValuePair(params, key, value)
}
\arguments{
\item{params}{Any parameter pertaining to a specific web method invocation.}

\item{key}{The key to add to the params variable.}

\item{value}{The value to add to the params variable.}
}
\description{
Function is used to add the key and value to the params that will be sent on
every call to OpenSecrets.org; an exception will be thrown if either of
these parameters is null.
}

