% Generated by roxygen2 (4.0.1): do not edit by hand
\name{AddMethod}
\alias{AddMethod}
\title{Function is used to set the OpenSecrets.org API method being called.}
\usage{
AddMethod(params, method)
}
\arguments{
\item{params}{Any parameter pertaining to a specific web method invocation.}

\item{method}{The web method being invoked -- see
\href{https://www.opensecrets.org/resources/create/api_doc.php}{here}.}
}
\description{
Function is used to set the OpenSecrets.org API method being called.
}

