% Generated by roxygen2 (4.0.1): do not edit by hand
\name{AddOutputAsJSON}
\alias{AddOutputAsJSON}
\title{Function is used to set the output parameter to JSON.}
\usage{
AddOutputAsJSON(params)
}
\arguments{
\item{params}{Any parameter pertaining to a specific web method invocation.}
}
\description{
Function is used to set the output parameter to JSON.
}

