% Generated by roxygen2 (4.0.1): do not edit by hand
\name{DoGet}
\alias{DoGet}
\title{Function performs the request for data from the OpenSecrets.org website and
returns the result as a data frame.}
\usage{
DoGet(method, params)
}
\arguments{
\item{method}{The web method being invoked -- see
\href{https://www.opensecrets.org/resources/create/api_doc.php}{here}.}

\item{params}{Any parameter pertaining to a specific web method invocation.}
}
\description{
Function performs the request for data from the OpenSecrets.org website and
returns the result as a data frame.
}

