% Generated by roxygen2 (4.0.1): do not edit by hand
\name{GetCandContribData}
\alias{GetCandContribData}
\title{Provides the top organizations contributing to specified politician.}
\usage{
GetCandContribData(params)
}
\arguments{
\item{params}{Any parameter accepted by this web service call -- see
\href{https://www.opensecrets.org/api/?output=doc&method=candContrib}{here}.}
}
\description{
Provides the top organizations contributing to specified politician.
}
\examples{
\dontrun{
 SetAPIKey ("ENTER YOUR PRIVATE API KEY HERE.")
 params <- list (cid="N00007360", cycle="2012")
 tryCatch(
     candContribData <- GetCandContribData (params),
     error =
         function (e) {
             print (
                 paste (
                     "An exception was thrown -- details follow: ",
                     e,
                     sep=""
                 )
             )
         }
     )
 }
}

