% Generated by roxygen2 (4.0.1): do not edit by hand
\name{GetCandIndByIndData}
\alias{GetCandIndByIndData}
\title{Provides total contributed to specified candidate from specified industry
for specified cycle.}
\usage{
GetCandIndByIndData(params)
}
\arguments{
\item{params}{Any parameter accepted by this web service call -- see
\href{https://www.opensecrets.org/api/?output=doc&method=candIndByInd}{here}.}
}
\description{
Provides total contributed to specified candidate from specified industry
for specified cycle.
}
\examples{
\dontrun{
 SetAPIKey ("ENTER YOUR PRIVATE API KEY HERE.")
 params <- list (cid="N00007360", cycle="2012", ind="K02")
 tryCatch(
     candIndByIndData <- GetCandIndByIndData (params),
     error =
         function (e) {
             print (
                 paste (
                     "An exception was thrown -- details follow: ",
                     e,
                     sep=""
                 )
             )
         }
     )
 }
}

