% Generated by roxygen2 (4.0.1): do not edit by hand
\name{GetCandSummaryData}
\alias{GetCandSummaryData}
\title{Provides summary fundraising information for a specified politician.}
\usage{
GetCandSummaryData(params)
}
\arguments{
\item{params}{Any parameter accepted by this web service call -- see
\href{https://www.opensecrets.org/api/?output=doc&method=candSummary}{here}.}
}
\description{
Provides summary fundraising information for a specified politician.
}
\examples{
\dontrun{
 SetAPIKey ("ENTER YOUR PRIVATE API KEY HERE.")
 params <- list (cid="N00007360", cycle="2012")
 tryCatch(
     candSummaryData <- GetCandSummaryData (params),
     error =
         function (e) {
             print (
                 paste (
                     "An exception was thrown -- details follow: ",
                     e,
                     sep=""
                 )
             )
         }
     )
 }
}

