% Generated by roxygen2 (4.0.1): do not edit by hand
\name{GetCongCmteIndusData}
\alias{GetCongCmteIndusData}
\title{Provides summary fundraising information for a specific committee, industry
and congress number.}
\usage{
GetCongCmteIndusData(params)
}
\arguments{
\item{params}{Any parameter accepted by this web service call -- see
\href{https://www.opensecrets.org/api/?output=doc&method=congCmteIndus}{here}.}
}
\description{
Provides summary fundraising information for a specific committee, industry
and congress number.
}
\examples{
\dontrun{
 SetAPIKey ("ENTER YOUR PRIVATE API KEY HERE.")
 params <- list (congno="112", indus="F10", cmte="HARM")
 tryCatch(
     congCmteIndusData <- GetCongCmteIndusData (params),
     error =
         function (e) {
             print (
                 paste (
                     "An exception was thrown -- details follow: ",
                     e,
                     sep=""
                 )
             )
         }
     )
 }
}

