% Generated by roxygen2 (4.0.1): do not edit by hand
\name{GetLegislatorsData}
\alias{GetLegislatorsData}
\title{Provides a list of legislators and associated attributes.}
\usage{
GetLegislatorsData(params)
}
\arguments{
\item{params}{Any parameter accepted by this web service call -- see
\href{https://www.opensecrets.org/api/?output=doc&method=getLegislators}{here}.}
}
\description{
Provides a list of legislators and associated attributes.
}
\examples{
\dontrun{
 SetAPIKey ("ENTER YOUR PRIVATE API KEY HERE.")
 params <- list (id="NJ")
 tryCatch(
     legislatorsData <- GetLegislatorsData (params),
     error =
         function (e) {
             print (
                 paste (
                     "An exception was thrown -- details follow: ",
                     e,
                     sep=""
                 )
             )
         }
     )
 }
}

