% Generated by roxygen2 (4.0.1): do not edit by hand
\name{GetMemPFDprofileData}
\alias{GetMemPFDprofileData}
\title{Returns PFD information for a member of Congress.}
\usage{
GetMemPFDprofileData(params)
}
\arguments{
\item{params}{Any parameter accepted by this web service call -- see
\href{https://www.opensecrets.org/api/?output=doc&method=memPFDprofile}{here}.}
}
\description{
Returns PFD information for a member of Congress.
}
\examples{
\dontrun{
 SetAPIKey ("ENTER YOUR PRIVATE API KEY HERE.")
 params <- list (year="2010", cid="N00007360")
 tryCatch(
     memPFDprofileData <- GetMemPFDprofileData (params),
     error =
         function (e) {
             print (
                 paste (
                     "An exception was thrown -- details follow: ",
                     e,
                     sep=""
                 )
             )
         }
     )
 }
}

