% Generated by roxygen2 (4.0.1): do not edit by hand
\name{GetOrgsData}
\alias{GetOrgsData}
\title{Provides organization data.}
\usage{
GetOrgsData(params)
}
\arguments{
\item{params}{Any parameter accepted by this web service call -- see
\href{https://www.opensecrets.org/api/?output=doc&method=getOrgs}{here}.}
}
\description{
Provides organization data.
}
\examples{
\dontrun{
 SetAPIKey ("ENTER YOUR PRIVATE API KEY HERE.")
 params <- list (org="Microsoft")
 tryCatch(
     orgsData <- GetOrgsData (params),
     error =
         function (e) {
             print (
                 paste (
                     "An exception was thrown -- details follow: ",
                     e,
                     sep=""
                 )
             )
         }
     )
 }
}

