% Generated by roxygen2 (4.0.1): do not edit by hand
\name{GetOrgSummaryData}
\alias{GetOrgSummaryData}
\title{Provides organization summary data.}
\usage{
GetOrgSummaryData(params)
}
\arguments{
\item{params}{Any parameter accepted by this web service call -- see
\href{https://www.opensecrets.org/api/?output=doc&method=orgSummary}{here}.}
}
\description{
Provides organization summary data.
}
\examples{
\dontrun{
 SetAPIKey ("ENTER YOUR PRIVATE API KEY HERE.")
 params <- list (id="123")
 tryCatch(
     orgSummaryData <- GetOrgSummaryData (params),
     error =
         function (e) {
             print (
                 paste (
                     "An exception was thrown -- details follow: ",
                     e,
                     sep=""
                 )
             )
         }
     )
}
}

