% Generated by roxygen2 (4.0.1): do not edit by hand
\name{SetAPIKey}
\alias{SetAPIKey}
\title{Allows the user to set the OpenSecrets.org API key once thereby removing the
need to pass it in for each function call in this package.}
\usage{
SetAPIKey(apiKey)
}
\arguments{
\item{apiKey}{The OpenSecrets.org API key.}
}
\description{
Allows the user to set the OpenSecrets.org API key once thereby removing the
need to pass it in for each function call in this package.
}
\examples{
SetAPIKey ("Example API Key")
}

